requirements.txt
# Core JAX/Flax dependencies
jax>=0.4.0
flax>=0.8.0
optax>=0.1.0

# Data and ML dependencies
tensorflow>=2.13.0
tensorflow-datasets>=4.9.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization dependencies
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration dependencies
PyYAML>=6.0

# Additional utilities
tqdm>=4.65.0